requirements.txt
wikipedia
chatbotAI
googletrans==4.0.0rc1
discord.py
akinator.py
Pillow
flask

aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
Deprecated==1.2.13
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==2.5
jwcrypto==1.0
multidict==5.2.0
Naked==0.1.31
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pyparsing==3.0.4
python-firebase==1.2
python-jwt==3.3.1
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
sseclient==0.0.27
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.3
yarl==1.7.2